apiserver: Remove dead code
